python/vpc-ec2-webserver/requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_ec2
aws_cdk.aws_iam
